completed all class templates

+++ New/Promotion.java
@@ -1,5 +1,5 @@
 package MacDonalds.New;
 
 public interface Promotion {
-    public int getDiscount();
+    public float getDiscount();
 }
